Refactor schemas and update payment service integration
