+++ .github/workflows/deploy.yml
@@ -13,28 +13,23 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v4
 
-    - name: Install pnpm
-      uses: pnpm/action-setup@v4
-      with:
-        version: latest
-
     - name: Setup Node.js
       uses: actions/setup-node@v4
       with:
         node-version: '20'
-        cache: 'pnpm'
+        cache: 'npm'
 
     - name: Install dependencies
-      run: pnpm install --frozen-lockfile
+      run: npm ci
 
     - name: Type check
-      run: pnpm run type-check || echo "TypeScript check skipped (no type-check script)"
+      run: npm run type-check || echo "TypeScript check skipped (no type-check script)"
 
     - name: Lint
-      run: pnpm run lint || echo "Lint check skipped (no lint script)"
+      run: npm run lint || echo "Lint check skipped (no lint script)"
 
     - name: Build frontend
-      run: pnpm run build
+      run: npm run build
 
     - name: Install Wrangler
       run: npm install -g wrangler
