+++ dev.vars.example
@@ -1,0 +1,11 @@
+# Local development environment variables (Worker Secrets Only)
+# Copy this file to .dev.vars and fill in your actual values
+
+# API Keys and Tokens (SENSITIVE)
+BLAWBY_API_TOKEN=your_blawby_api_token_here
+LAWYER_SEARCH_API_KEY=your_lawyer_search_api_key_here
+CLOUDFLARE_API_TOKEN=your_cloudflare_api_token_here
+RESEND_API_KEY=your_resend_api_key_here
+
+# Account IDs (SENSITIVE)
+CLOUDFLARE_ACCOUNT_ID=your_cloudflare_account_id_here
