--- tests/stubs/worker_threads.cjs
@@ -1,5 +1,0 @@
-module.exports = new Proxy({}, {
-  get() {
-    throw new Error('node:worker_threads is not available in workerd test environment');
-  }
-});
