+++ worker/schemas/index.ts
@@ -1,0 +1,13 @@
+// Centralized schema exports
+
+// Validation schemas (Zod)
+export * from './validation';
+
+// Type definitions
+export * from './jurisdictionConfig';
+export * from './payment';
+export * from './lawyer';
+export * from './legal';
+
+// Re-export common types from types.ts
+export type { Env, HttpError } from '../types';
